linkCardのpostメソッドをreduxで実装
reduxからactionを受け取り、コンポーネントはアクションを受け取り利用するだけとした。
