+++ client/package.json
@@ -5,9 +5,11 @@
   "dependencies": {
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
+    "react-redux": "^7.1.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1",
     "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
     "superagent": "^5.1.0"
   },
   "scripts": {
