+++ client/src/mainPage/PostCard.js
@@ -1,6 +1,10 @@
 import React from "react";
+import { connect } from "react-redux";
+import { bindActionCreators } from "redux";
 import { withRouter, Link } from "react-router-dom";
 import request from "superagent";
+
+import { addFetchLinkCard } from "../reducer/modules/linkCards";
 
 class PostCard extends React.Component {
   constructor(props) {
@@ -21,19 +25,11 @@ class PostCard extends React.Component {
   }
 
   post() {
-    request
-      .post("/api/createLink")
-      .send({
-        title: this.state.title,
-        comment: this.state.comment,
-        url: this.state.url
-      })
-      .end((err, data) => {
-        if (err) {
-          console.error(err);
-          return;
-        }
-      });
+    this.props.addFetchLinkCard({
+      title: this.state.title,
+      comment: this.state.comment,
+      url: this.state.url
+    });
     this.props.history.push("/");
   }
 
@@ -66,15 +62,6 @@ class PostCard extends React.Component {
       sendTitleCheck = "ok";
     }
 
-    // switch(alertTitleType) {
-    //   case "long":
-    //     msgTitle = <p style={displayStyle}>*20文字以内</p>
-    //   case "none":
-    //     msgTitle = <p></p>
-    //   case "short":
-    //     msgTitle = <p style={displayStyle}>*必須</p>
-    // }
-
     if (this.state.url === "") {
       msgUrl = <p style={displayStyle}>*必須</p>;
       sendUrlCheck = "ng";
@@ -85,13 +72,6 @@ class PostCard extends React.Component {
       msgUrl = <p style={displayStyle}>リンクを貼ってください</p>;
       sendUrlCheck = "ng";
     }
-
-    // switch(alertUrlType) {
-    //   case true:
-    //     msgUrl = <p style={displayStyle}>*必須</p>
-    //   case false:
-    //     msgUrl = <p></p>
-    // }
 
     if (
       this.state.title &&
@@ -157,6 +137,10 @@ class PostCard extends React.Component {
   }
 }
 
+const mapDispatchToProps = dispatch => {
+  return bindActionCreators({ addFetchLinkCard }, dispatch);
+};
+
 const styles = {
   Link: {
     display: "block",
@@ -168,4 +152,9 @@ const styles = {
   }
 };
 
-export default withRouter(PostCard);
+export default withRouter(
+  connect(
+    null,
+    mapDispatchToProps
+  )(PostCard)
+);
